+++ lib/templates/gametemplate.js
@@ -97,6 +97,7 @@ class GameTemplate extends ModTemplate {
     this.can_bet = 1; // if it is possible to win this game
 
     this.request_no_interrupts = true;
+    this.acknowledge_text = "I understand...";
 
     this.confirm_moves = 0;
     this.confirm_this_move = 0;
@@ -3282,7 +3283,7 @@ class GameTemplate extends ModTemplate {
   }
 
   playerAcknowledgeNotice(msg, mycallback) {
-    let html = `<ul><li class="textchoice acknowledge" id="confirmit">I understand...</li></ul>`;
+    let html = `<ul><li class="textchoice acknowledge" id="confirmit">${this.acknowledge_text}</li></ul>`;
 
     try {
       this.updateStatusWithOptions(msg, html);
